backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
redis==5.0.8
slowapi==0.1.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.2
tenacity==9.0.0
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.2
pytest-asyncio==0.23.8
requests==2.32.3



